Modernize apt installation instructions and improve security
Modern Debian-derived distributions will issue a deprecation warning for
the usage of apt-key add as listed here.  Additionally, by using that
apt-key command, the given key is recognized as valid for all
repositories in /etc/apt/sources.list* -- not just this one.

By using the more modern mechanism, we can avoid that warning and also
import this key in such a way that it is only valid for this repository.

Other projects are also using this method; for instance:

https://apt.syncthing.net/

This post explains the security risks of apt-key:
https://blog.cloudflare.com/dont-use-apt-key/